requirements.txt
pyspark==3.3.0
psycopg2-binary~=2.9.4
numpy~=1.23.4